Add the option to export just one release
